requirements.txt
pypng==0.0.20
tqdm==4.60.0
numpy==1.20.3
Pillow==8.2.0